+++ src/uproot/behaviors/TBranch.py
@@ -152,7 +152,7 @@ def iterate(
 
     Options (type; default):
 
-    * handler (:doc:`uproot.source.chunk.Source` class; `uproot.source.fsspec.FSSpecSource`)
+    * handler (:doc:`uproot.source.chunk.Source` class; `None`)
     * timeout (float for HTTP, int for XRootD; 30)
     * max_num_elements (None or int; None)
     * num_workers (int; 1)
@@ -321,7 +321,7 @@ def concatenate(
 
     Options (type; default):
 
-    * handler (:doc:`uproot.source.chunk.Source` class; `uproot.source.fsspec.FSSpecSource`)
+    * handler (:doc:`uproot.source.chunk.Source` class; `None`)
     * timeout (float for HTTP, int for XRootD; 30)
     * max_num_elements (None or int; None)
     * num_workers (int; 1)
